BUG 2847, 2848, 2841, 2844: Various fixes * Don't allow inline editing of column names in views * Fix sidebar widths so pieces don't wrap * Change datepicker arrows, since the solid arrows didn't appear in IE * Fix various issues with group views: ** Editing invalid grouped view was throwing an error ** Above wasn't being marked as valid, since the message wasn't remove in Dataset when getting the view back ** Don't initially request rows for an invalid view (in fact, don't load grid at all) ** After editing an invalid view or when saving a regular grouped view after changing the grouped column, the newly-chosen column was hidden.  So make sure on saving a dataset/view that the visible columns are set correctly * Clean up some misc params in datasetGrid that weren't used anymore
QA:
* Make sure bugs are fixed

+++ public/javascripts/screens/dataset-show.js
@@ -24,6 +24,7 @@ blist.datasetPage.updateValidView = function()
 {
     $('.invalidView').removeClass('invalidView');
     datasetPageNS.sidebar.updateEnabledSubPanes();
+    datasetPageNS.initGrid();
 };
 
 blist.datasetPage.hidePageRenderType = function()
@@ -55,22 +56,20 @@ blist.datasetPage.showPageRenderType = function()
 
 blist.datasetPage.initGrid = function()
 {
-    if (datasetPageNS.gridInitialized || !blist.dataset.isGrid()) { return; }
+    if (datasetPageNS.gridInitialized || !blist.dataset.isGrid() ||
+        !blist.dataset.valid) { return; }
 
     datasetPageNS.$dataGrid
         .datasetGrid({view: blist.dataset,
             columnDeleteEnabled: blist.dataset.type == 'blist' &&
                 blist.dataset.hasRight('remove_column'),
             columnPropertiesEnabled: true,
-            columnNameEdit: blist.dataset.hasRight('update_view'),
+            columnNameEdit: blist.dataset.type == 'blist' &&
+                blist.dataset.hasRight('update_view'),
             showAddColumns: blist.dataset.type == 'blist' &&
                 blist.dataset.hasRight('add_column'),
             accessType: 'WEBSITE', manualResize: true, showRowHandle: true,
-            clearTempViewCallback: datasetPageNS.clearTempView,
-            setTempViewCallback: datasetPageNS.setTempView,
             filterForm: '#searchForm', clearFilterItem: '#searchForm .clearSearch',
-            isInvalid: !blist.dataset.valid,
-            validViewCallback: datasetPageNS.updateValidView,
             addColumnCallback: function(parId)
             {
                 datasetPageNS.sidebar.addPane('edit.addColumn', {parentId: parId});
